* commit 'e514f20c713c7ccaf1bc3946bb45a503dfe03406':
  calculation.cc : revised fix for size_type logging (GCC casts it to size_t, and for that we can simply print C99 %zu)
  Makefile.am : added the new src/db/calculation.h to EXTRA_DIST to pass distcheck (typo fix)
  Makefile.am : added the new db/calculation.h to EXTRA_DIST to pass distcheck
  ci-test-netmon.sh and vte-netmon.sh : stop the daemons before parsing output, in particular to flush buffered output into the dsh_file
  ci-test-netmon.sh and vte-netmon.sh : remove DSH_FILE if DEBUG is not in place; report if there was error reading one and abort visibly
  BIOS-747 : ci-setup-test-machine.sh : separate mk-build-deps and apt-get/dpkg install, because when combined it tends to remove the generated .deb file
  BIOS-747 : autogen.sh : separate mk-build-deps and apt-get/dpkg install, because when combined it tends to remove the generated .deb file
  agent-tpower.h : small typo fix
  Unified agent naming pattern: renamed agent-alert.cc and .h
  Unified agent naming pattern: renamed agent-alert-smtp.cc and .h, and a small typo fix
  Unified agent naming pattern: renamed agent-autoconfig.cc and .h
  Unified agent naming pattern: renamed agent-inventory.cc
  Unified agent naming pattern: renamed agent-cm.cc
  BIOS-957 : bios-db-init.service.in added, to separate actual dependency of services on an existing schema rather than a database service per se
  Unified agent naming pattern: renamed agent-nut.cc and .h
  Unified agent naming pattern: renamed agent-dbstore.cc
  Unified agent naming pattern: renamed agent-netmon.c
  agent-tpower.cc : do have a return value on exception
  calculation.cc : fixed map size() logging to a certain uint64 type

+++ src/db/calculation.cc
@@ -331,7 +331,7 @@ reply_t
         true,
         measurements,
         unit);
-    log_debug (" selected %" PRIu64, (long unsigned int)(measurements.size()) );
+    log_debug (" selected %zu", measurements.size() );
     LOG_END;
     return ret;
 }
